Add files via upload
This was deployed before two years ago on Nov 10, 2022 at 4:06 PM to be specific. I started a JavaScript course on Scrimba but I never got to finish it. Today whilst writing this it is February the 9th, 2024, 6:34 PM to be specfic and I will finish this course and get commited ;)